+++ src/torchmetrics/audio/srmr.py
@@ -53,6 +53,12 @@ class SpeechReverberationModulationEnergyRatio(Metric):
         Either install as ``pip install torchmetrics[audio]`` or ``pip install torchaudio``
         and ``pip install git+https://github.com/detly/gammatone``.
 
+    .. note::
+        This implementation is experimental, and might not be consistent with the matlab
+        implementation `SRMRToolbox`_, especially the fast implementation.
+        The slow versions, a) fast=False, norm=False, max_cf=128, b) fast=False, norm=True, max_cf=30, have
+        a relatively small inconsistence.
+
     Args:
         fs: the sampling rate
         n_cochlear_filters: Number of filters in the acoustic filterbank
@@ -93,7 +99,7 @@ class SpeechReverberationModulationEnergyRatio(Metric):
         n_cochlear_filters: int = 23,
         low_freq: float = 125,
         min_cf: float = 4,
-        max_cf: Optional[float] = 128,
+        max_cf: Optional[float] = None,
         norm: bool = False,
         fast: bool = False,
         **kwargs: Any,